+++ luapress/config.lua
@@ -1,3 +1,3 @@
 return {
-    version = '3.0'
+    version = '3.1'
 }