+++ CHANGELOG.md
@@ -11,8 +11,11 @@ The rest:
 + Add `config.force_index_page` to create index form page even when posts exist
 + Add `config.sticky_page` to have a page appear at the top of index
 + Add `config.archive_title` to change the title of the archive page
-+ Add default config, so local config can be sparse
 + Add support for multiple (url, build_dir) environments
++ Add default config, so local config can be sparse
++ Add ID attribute to `<hX>` tags in markdown lib
++ Fix escaping for post & gallery RSS content
++ Rename `press.lua` -> `luapress.lua`
 + Remove `alt_getopt` and handle args manually
 
 # 2.1.1